+++ lib/gitHook.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 /*
  * Copyright © 2018 Atomist, Inc.
  *
@@ -16,12 +14,24 @@
  * limitations under the License.
  */
 
-process.env.ATOMIST_DISABLE_LOGGING = "true";
+import {
+    runOnGitHook,
+} from "@atomist/sdm-local";
 
-import { runOnGitHook } from "@atomist/sdm-local";
+import * as print from "./print";
 
-process.stderr.write(`githook: [WARN] The githook script is deprecated.  Use 'atomist git-hook'.`);
-runOnGitHook(process.argv).catch(err => {
-    process.stderr.write(err.message + "\n");
-    process.exit(1);
-});
+/**
+ * Generate git-info.json for automation client.
+ *
+ * @param opts see GitOptions
+ * @return integer return value
+ */
+export async function gitHook(args: string[]): Promise<number> {
+    try {
+        await runOnGitHook(args);
+    } catch (e) {
+        print.error(`Failed to process Git hook: ${e.message}`);
+        return 1;
+    }
+    return 0;
+}